Crate imprint

Provides the ability to imprint values at the type level, enabling compile-time validation of values that only exist at run time.

Heavily inspired by Richard Eisenberg's singletons library, Edward Kmett's reflection and eq libraries, as well as Gankro's sound unchecked indexing approach.

Modules

arith
As with Val, the soundness and usefulness of these traits rely on the type contained. In particular, in order to enable sensible use of these relations:

ix

Structs

Exists
An object with an existentially quantified lifetime.

IdF
Identity function for types.

TyEq
Propositional equality between types.

Val
A value imprinted at the type level.

ValF
Allows Val to be parameterized by its lifetime parameter. Here, F can be any arbitrary type function.

Traits

TyFn
Used to define type-level functions.

TyFnL
Used to define type-level functions with existential parameters, intended for use with Exists.

Value
Any type that implements Value represents a promoted value.

Functions

imprint
Imprint the type of an object with its own value.

Type Definitions

PhantomInvariantLifetime
Like PhantomData but ensures that 'a is always invariant.

PhantomInvariantType
Like PhantomData but ensures that T is always invariant.