Struct imprint::Val

pub struct Val<'x, T> { /* fields omitted */ }

A value imprinted at the type level.

A Val<'x, T> value contains an instance of T as well as a marker 'x that reflects the value of that instance at the type level. This provides a type-safe mechanism to constrain values even if their actual values are not known at compile time.

Val can be constructed using either imprint(...) or Default::default().

The underlying value can be obtained either by dererefencing or by calling .value().

Properties

The notion of "value" is determined by the equivalence relation formed by Eq, or the partial equivalence relation formed by PartialEq.

We expect the value of T to be immutable through &T. Otherwise, the properties in this section would not hold. Therefore, Val is not very useful for types with interior mutability like Cell or RefCell. Moreover, keep in mind that any unsafe code can violate these properties as well.

If T forms an equivalence relation, then for every marker 'x, the type Val<'x, T> contains precisely one value, and each value corresponds to a unique 'x. Hence, Val<'x, T> may be considered a singleton type (unrelated to "singletons" in OOP).

On the other hand, if T forms a partial equivalence relation, then for every marker 'x, the type Val<'x, T> contains either a single identifiable value (for which equality is reflexive), or a single unidentifiable value (for which equality is nonreflexive), and each identifiable value corresponds to a unique marker 'x.

Methods

impl<'x, T> Val<'x, T>

fn as_val_ref<'a>(&'a self) -> Val<'x, &'a T>

unsafe fn known(value: T) -> Self

impl<'a, 'x, T: Clone> Val<'x, &'a T>

fn cloned(self) -> Val<'x, T>

impl<'x, T: PartialEq> Val<'x, T>

fn eq<'y>(&self, other: &Val<'y, T>) -> Option<TyEq<Self, Val<'y, T>>>

Checks whether two values are equal. If they are, evidence of their equality is returned.