Struct imprint::ValF

pub struct ValF<F: ?Sized, T>(_, _);

Allows Val to be parameterized by its lifetime parameter. Here, F can be any arbitrary type function.

Here, F is required to be parametric, but atm we have no way to enforce that. The higher-ranked trait bound will ensure parmetricity for now, but that may change in the future.