Trait imprint::Value

pub unsafe trait Value {
    type Value;
    fn value(self) -> Self::Value;
}

Any type that implements Value represents a promoted value.

Unsafe assumptions

This trait is unsafe: any type that implements Value must effectively have only one possible value. Failing to abide by this will lead to unsafety. (E.g. if one has both Equal<Good1, Bad> and Equal<Bad, Good2>, then by transitivity we have Equal<Good1, Good2>, but Bad could be multivalued!)

Associated Types

type Value

The demoted type.

Required Methods

fn value(self) -> Self::Value

Extracts the value.